fixed a menu and updated logger

+++ src/commands/message/delete.ts
@@ -6,10 +6,14 @@ export default new MessageContextMenu({
     .setName("delete")
     .setType(ApplicationCommandType.Message),
   run: async ({ client, interaction }) => {
-    await interaction.targetMessage.delete();
-    interaction.reply({
-      content: `Message has been deleted!`,
-      ephemeral: true,
-    });
+    if (interaction.memberPermissions.has("ManageMessages")) {
+      await interaction.targetMessage.delete();
+      interaction.reply({
+        content: `Message has been deleted!`,
+        ephemeral: true,
+      });
+    } else {
+     await interaction.reply({ content: `You are missing permissions to delete messages`})
+    }
   },
 });
